fixed 4553c6a from https://gitee.com/hang_jian/arkui_ace_engine/pulls/26600 修改告警问题2

+++ frameworks/core/components_ng/pattern/rich_editor/rich_editor_pattern.cpp
@@ -5161,7 +5161,7 @@ bool RichEditorPattern::NeedAiAnalysis(
         return false;
     }
 
-    if (pos == content.length()) {
+    if (pos == static_cast<int32_t>(content.length())) {
         return false;
     }
